Sync to published messages
This commit uses the ``client_token`` mechanism of AWS IoT service to
synchronize to replies from upstream.

The client class becomes relatively complicated after the change, which makes
unit tests hard and flaky to set up. The unit tests for IoT server interaction
are disabled in favor of manual testing for now.