Fix ``Client`` only subscribing to updates in ``__aenter()__``
While it's recommended to use the context manager, it should not be
mandated. Client is now ready to be used without entering context.

Also remove the redundant ``loop.call_soon_threadsafe()`` call when handling
callbacks from MQTT updates. The raw MQTT callback already ensures that it's
called in event loop.